chore: redistribute code from color mode context
