+++ packages/apps/human-app/frontend/src/main.tsx
@@ -16,7 +16,7 @@ import '@fontsource/inter/800.css';
 import { WalletConnectProvider } from '@/shared/contexts/wallet-connect';
 import { Web3AuthProvider } from '@/modules/auth-web3/context/web3-auth-context';
 import { JWTExpirationCheck } from '@/shared/contexts/jwt-expiration-check';
-import { ColorModeProvider } from '@/shared/contexts/color-mode-context';
+import { ColorModeProvider } from '@/shared/contexts/color-mode';
 import { HomePageStateProvider } from '@/shared/contexts/homepage-state';
 import { RegisteredOraclesProvider } from '@/shared/contexts/registered-oracles';
 import { NotificationProvider } from '@/shared/providers/notifications-provider';
